Add files via upload
A program that checks whether numbers are factors of 3 and 5. If they are, the program assigns the word Fizz to factors of 3. If a numbers is a factor of 5, than the word "Buzz" is displayed instead of the numerical value and if factors of both 3 and 5 are found, the word "FizzBuzz" is displayed rather than  number itself.