Add files via upload
The program sorts out an array that is randomized. It uses two loops, an inner loop and a outer loop. The inner loop is used to iterate through out the entirety of the loop length per the current index. The out loop is assigned to begin at the zero index but the outer loop begins at the first index. An if statement is used to compare both values with each other per iteration of the loops. If the value of the predecessor is greater than of the successor, the values are swapped so that the greatest value comes after the lesser value.