Add files via upload
The program consists of two loops. An inner loop and a outer loop. The first loop begins at the index of the array and the second loop begins just after the 0 index. 
and if statement is used to compare the values of the array. If the predecessor is greater than the successor than the values are swapped ensuring that the array is sorted in an ascending order